// frontend/ticket-booking-web/src/app/api/model/version-info-response.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VersionInfoResponse { 
    version: string;
    commit: string;
    runtimeVersion: string;
    runtimeMajor: number;
    buildDate: string | null;
    environment: string;
}